Core: factor out time.sleep usage
We can use the fact that Watchdog observers are threads to our benefit here and use join instead of time.sleep().

Note: In Python 3, thie `While True` wrapper and the timeout passed to join aren't necessary. KeyboardInterrupts give control back to the main thread. That isn't the case in Python < 3, so we still need to keep that annoying wrapper.

 Making the `observer` thread a demon also means we no longer need to explicitly wait for it to shut down.